Move client logic from main() to client::mod.rs
