+++ src/main.rs
@@ -14,11 +14,8 @@ mod server;
 mod util;
 mod view;
 
-use bincode::serde::deserialize_from;
 use clap::AppSettings;
-use model::types::{Board, SubField, Mode};
-use net::types::{self, MessageType};
-use std::net::{Shutdown, TcpStream};
+use model::types::{Board, Mode};
 use term_painter::ToStyle;
 use term_painter::Color::*;
 
@@ -128,20 +125,11 @@ fn main() {
                 port,
             );
 
-            // create client instance and connect to server
-            let mut connection = TcpStream::connect((ip, port)).unwrap();
-            // add CTRL+C system hook, so that connection partner is informed about disconnect
-            let client_conn_clone = connection.try_clone().unwrap();
-            ctrlc::set_handler(move || {
-                net::send(&mut client_conn_clone.try_clone().unwrap(), MessageType::Quit);
-                client_conn_clone.shutdown(Shutdown::Both).expect("shutdown call failed");
-            }).expect("Error setting Ctrl+C handler");
-
             let board = client_args.value_of("board")
                 .map(|b| util::read_extern_board(b))
                 .unwrap_or(Board::init());
 
-            let mut client = ::model::types::Player {
+            let client = ::model::types::Player {
                 own_board: board.clone(),
                 op_board: Board::init(),
                 player_type: ::model::types::PlayerType::Human,
@@ -149,130 +137,8 @@ fn main() {
                 name: name.to_string(),
             };
 
-            let mut host_name = "SERVER".to_string();
-            loop {
-                //let mut buffer = Vec::new();
-                //let msg = client_connection.read_to_end(&mut buffer);
-                //println!("{}", str::from_utf8(&buffer).unwrap());
-                let recv: Result<types::MessageType, _> =
-                    deserialize_from(&mut connection, bincode::SizeLimit::Infinite);
-                match recv {
-                    Ok(received) => {
-                        // process_message(received);
-                        match received {
-                            MessageType::Welcome(msg, host) => {
-                                Yellow.with(|| println!("{}", (msg)));
-                                host_name = host;
-                                net::send(&mut connection, MessageType::Login(name.to_string()));
-                            },
-                            MessageType::Ping => {
-                                net::send(&mut connection, MessageType::Ping);
-                            },
-                            MessageType::Quit => {
-                                println!("Server ended the connection.");
-                                break;
-                            },
-                            MessageType::RequestCoord => {
-                                Yellow.with(|| print!("It's your turn! "));
-                                // send coordinate to shoot
-                                let mut coord;
-                                loop {
-                                    Yellow.with(|| println!("Please enter a valid coordinate: "));
-                                    coord = util::read_string();
-                                    if ::model::valid_coordinate(&coord) {
-                                        break;
-                                    }
-                                    Red.with(|| print!("Invalid coordinate! "));
-                                }
-
-                                net::send(&mut connection, MessageType::Shoot(coord));
-
-                                // receive updated opponent board
-                                let result: Result<types::MessageType, _> =
-                                    deserialize_from(&mut connection, bincode::SizeLimit::Infinite);
-                                match result {
-                                    Ok(res) => {
-                                        match res {
-                                            MessageType::Hit(id) => {
-                                                Green.with(|| println!("Hit!"));
-                                                client.op_board[id] = SubField::Hit;
-                                            }
-                                            MessageType::Miss(id) => {
-                                                Blue.with(|| println!("Miss!"));
-                                                client.op_board[id] = SubField::Miss;
-                                            }
-                                            _ => {}
-                                        }
-                                    }
-                                    Err(_) => println!("Did not receive Hit or Miss message.")
-                                }
-                                model::print_boards(&client);
-                            }
-                            MessageType::RequestBoard => {
-                                model::print_boards(&client);
-
-                                if client.capacity == 0 {
-                                    match model::place_ships(&mut client) {
-                                        Ok(()) => {},
-                                        Err(_) => println!("Failure on placing ships.")
-                                    }
-                                }
-
-                                // send board
-                                net::send(
-                                    &mut connection,
-                                    MessageType::Board(client.own_board.clone())
-                                );
-                            }
-                            MessageType::Text(t) => {
-                                Cyan.with(|| println!("{}", t));
-                            }
-                            MessageType::TurnHost => {
-                                println!(
-                                    "{} {} {}",
-                                    Cyan.paint("Wait for"),
-                                    Yellow.paint(&host_name),
-                                    Cyan.paint("to finish turn!"),
-                                );
-
-                                let result: Result<types::MessageType, _> =
-                                    deserialize_from(&mut connection, bincode::SizeLimit::Infinite);
-                                match result {
-                                    Ok(res) => {
-                                        match res {
-                                            MessageType::Hit(id) => {
-                                                client.own_board[id] = SubField::Hit;
-                                            }
-                                            MessageType::Miss(id) => {
-                                                client.own_board[id] = SubField::Miss;
-                                            }
-                                            _ => {}
-                                        }
-                                    }
-                                    Err(_) => println!("Did not receive Hit or Miss message.")
-                                }
-                                model::print_boards(&client);
-                            }
-                            MessageType::Unexpected => {
-                                Red.with(|| println!("Handshake done wrong!"));
-                            }
-                            MessageType::Lost => {
-                                Yellow.with(|| println!("You lost the game :("));
-                            }
-                            MessageType::Won => {
-                                Yellow.with(|| println!("Congratulations, you won the game!"));
-                            }
-                            _ => {
-                                Red.with(|| println!("Received unexpected packet"));
-                            }
-                        }
-                    },
-                    Err(_) => {
-                        Red.with(|| println!("Connection dropped..."));
-                        break;
-                    },
-                };
-            }
+            // connect to server
+            client::connect(client, ip, port);
         },
 
         ("single", Some(single_args)) => {
